PolyQuant/PolyScalping/PolyQuant/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["polyquant*"]

[project]
name = "polyquant"
version = "0.1.0"
description = "Data ingestion pipeline for Polymarket 15-minute Up/Down markets"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "PolyQuant Team"}
]
dependencies = [
    "requests>=2.31.0",
    "pandas>=2.0.0",
    "pyarrow>=12.0.0",
    "python-dateutil>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "types-requests>=2.31.0",
    "types-python-dateutil>=2.8.0",
]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
